Fix duplicate entries in the isolated_workloads list (#142)
Actual writes to the filesystem are asynchronous so add/remove
operations have no ordering guarantees.  This causes occasional leaking
of workload IDs.  Here we lazily remove any entries which are unknown to
the workload manager.  It is impossible for a workload to be isolated if
the workload manager isn't tracking it.