Yield to other threads when processing events.
When titus-isolate crashes, and comes up with a large backlog of
events, it can take a long time for it to process all the events.
System startup timer can expire in this period, causing a restart
loop.

This should yield the thread, and allow the systemd notification
to be processed when such a backlog occurs.