Make allocation requests declarative (#348)
Instead of the assign/free/rebalance imperative approach make
allocation requests describe the current state as layed out on a CPU,
and what workloads should be running on the CPU.

It is the CpuAllocator's responsibility to then do the right thing.  The
current design is still processing events from Docker serially, but we
can move to a model which batches those events once this new declarative
approach is adopted.