Ignore containers and events that are not main titus containers
As we add more containers alongside main containers,
we need to handle them properly. Certainly not try to
query the k8s api for them.

I'm not 100% sure if titus-isolate needs to actually
do anything with them? As they share the same pid namespace
as the main container and share its resources?

For now, this code ignores any container that does
not look like a Titus Task ID.